windows/requirements.txt
matplotlib==3.8.2
numpy==1.26.2
opencv_python==4.8.1.78
Pillow==10.1.0
Pillow==10.1.0
PyAutoGUI==0.9.54
pyperclip==1.8.2
Requests==2.31.0
